Begin using Web audio API to load audio
The Web Audio API has methods for loading data arrays and parsing WAV
files. Unfortunately it resamples into the browsers preferred sample
rate, which is typically 44100Hz. That is bad for making spectrograms
because all those extra samples slow down the FFT.

This commit actually contains two ways of dealing with this. The first
thing I tried was downsampling by an integer ratio to a rate close to but
above 8kHz. In the case of 44.1kHz, that was 8.82kHz. Integral
downsampling is much simpler than rational downsampling, as far as I can
see. It worked, though of course left a silent black band above the
original 4kHz Nyquist frequency.

But then I decided it wouldn't be too hard to just parse the WAV myself,
so I did that. It also worked and seems a better solution.

I have also remembered seeing somewhere that audiofile.js can deal with
wav data in list-like form, and that would also work.  But the wav
parsing function only turns out to be 30 lines long, and much of that is
dealing with unlikely events like 8-bit samples and multi-channel audio
(untested nevertheless).

The Web Audio API is a new thing for Firefox, though Chromium has had
it for a while.