Attempt simple_kiwi style detection of various birds
The theory is that the calls of kiwis and many other birds are at a
roughly fixed frequency and oscillate on and off at a certain rate. This
can be measured by recursively taking the spectral intensity of a
frequency band, as is done in the kokako project.

The results are not great (see images/simple-detector-wrong-1.png).
Something is not right.

This commit contains a whole lot of marginally related changes, such
as the RMS normalisation of the wav audio, and the experimental use of
a vorbis window instead of hann for the main histogram.