Allow files to be tagged as interesting
Interesting files contain things that seem likely to produce false
positives -- that is, an unusual sound that uses the call frequencies.
If an interesting file is an extract of a larger file (e.g. one minute
from a 15 minute recording), it could be productive to look further at
the rest of the interesting file.

This is quite a hack in terms of the database.