Use sound position when drawing playing line, not dead reckoning
Before the line would be advanced a pixel at a time at a supposedly
regular interval without reference to the sound position, which appeared
to work well over short stretches, but could lose its place by a second or
so over several minutes.

Now we ask the audio where it is, and work from there. To keep things
moderately smooth, we need to check more often.