extract-moreporks: work with JSON labels with intensity threshold
These labels are made by something in the kokako project, and have the
form of a file of JSON one-liners. The following, if you imagine it
all being on one line, would represent two calls in the file
"AI137_BIRM_120203_052707-09+60.wav":

       ["AI137_BIRM_120203_052707-09+60.wav",
        [0.25, 1.0, 3.894248141223105],
        [4.55, 5.35, 4.376498257003673]]

Each call has a start time, end time, and intensity score in that
order.